Add: #45783 - Hook alter menu / dynamic field manager service
